--- app/Validators/StoreFirstMessageValidator.ts
@@ -1,56 +1,0 @@
-import { schema, rules } from '@ioc:Adonis/Core/Validator'
-import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
-
-export default class StoreFirstMessageValidator {
-  constructor (protected ctx: HttpContextContract) {
-  }
-
-	/*
-	 * Define schema to validate the "shape", "type", "formatting" and "integrity" of data.
-	 *
-	 * For example:
-	 * 1. The username must be of data type string. But then also, it should
-	 *    not contain special characters or numbers.
-	 *    ```
-	 *     schema.string({}, [ rules.alpha() ])
-	 *    ```
-	 *
-	 * 2. The email must be of data type string, formatted as a valid
-	 *    email. But also, not used by any other user.
-	 *    ```
-	 *     schema.string({}, [
-	 *       rules.email(),
-	 *       rules.unique({ table: 'users', column: 'email' }),
-	 *     ])
-	 *    ```
-	 */
-  public schema = schema.create({
-    receiver_username: schema.string({}, [
-		rules.required()
-	]),
-	receiver_tag: schema.string({}, [
-		rules.required()
-	]),
-	content: schema.string({},[
-		rules.required()
-	])
-  })
-
-	/**
-	 * Custom messages for validation failures. You can make use of dot notation `(.)`
-	 * for targeting nested fields and array expressions `(*)` for targeting all
-	 * children of an array. For example:
-	 *
-	 * {
-	 *   'profile.username.required': 'Username is required',
-	 *   'scores.*.number': 'Define scores as valid numbers'
-	 * }
-	 *
-	 */
-  public messages = {
-	  'required': 'The {{ field }} field is required',
-	  'email': 'The {{ field }} is not a valid email', 
-	  'exists': 'The {{ field }} field does not exists on our db',  //⚠️ This can be insecure to let people just request our api and know if an email is registered to it
-	  'string': 'The {{ field }} field should be a string'
-  }
-}
